refactor: extract session recovery from api-core transport
